requirements/base.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
celery==5.3.6
django-celery-beat==2.6.0
redis==5.0.3
python-json-logger==2.0.7
dj-database-url==2.1.0
django-filter==24.1
Pillow==10.3.0
gunicorn==21.2.0
whitenoise[brotli]==6.6.0
